% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firesale-package.R
\docType{package}
\name{firesale-package}
\alias{firesale}
\alias{firesale-package}
\title{firesale: Data Store for 'fiery' web servers}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

}
\keyword{internal}
