% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firesale-package.R
\docType{package}
\name{firesale-package}
\alias{firesale}
\alias{firesale-package}
\title{firesale: Datastore for 'fiery' Web Servers}
\description{
Provides a persistent datastore for 'fiery' apps. The datastore is build on top of the 'storr' package and can thus be based on a variety of backends. The datastore contains both a global and session-scoped section.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/firesale}
  \item Report bugs at \url{https://github.com/thomasp85/firesale/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
